ST WRP re-organise C files
