+++ config/splat.pspeu.stwrp.yaml
@@ -41,19 +41,20 @@ segments:
     align: 128
     subalign: 8
     subsegments:
-      - [0x80, c, wrp_psp/80]
-      - [0x4D0, c, wrp_psp/4D0]
-      - [0x31F0, c, wrp_psp/31F0]
-      - [0x3C70, c, wrp_psp/3C70]
-      - [0x4100, c, wrp_psp/4100]
-      - [0x5600, c, wrp_psp/5600]
-      - [0x5AB0, c, wrp_psp/5AB0]
-      - [0x6BF0, c, wrp_psp/6BF0]
-      - [0x8580, c, wrp_psp/8580]
-      - [0x95D0, c, wrp_psp/95D0]
-      - [0xBDA0, c, wrp_psp/BDA0]
-      - [0xBFD8, c, wrp_psp/BFD8]
-      - [0xC3F0, c, wrp_psp/C3F0]
+      - [0x80, c, wrp_psp/create_entity]
+      - [0x1028, c, wrp_psp/e_stage_name]
+      - [0x1EA8, c, wrp_psp/e_red_door]
+      - [0x31F0, c, wrp_psp/st_update]
+      - [0x36E8, c, wrp_psp/e_particles]
+      - [0x4C78, c, wrp_psp/blit_char]
+      - [0x5308, c, wrp_psp/e_collect]
+      - [0x7448, c, wrp_psp/e_misc]
+      - [0x95D0, c, wrp_psp/collision]
+      - [0xB7A8, c, wrp_psp/popup]
+      - [0xBDA0, c, wrp_psp/e_room_fg]
+      - [0xBEB8, c, wrp_psp/st_common]
+      - [0xC9A0, c, wrp_psp/stage_name_new]
+      - [0xCEA0, c, wrp_psp/prim_helpers]
       - [0xD0E0, c, wrp_psp/st_init_wrp]
       - [0xD1B8, c, wrp/e_breakable]
       - [0xD338, c, wrp/st_debug]
