--- src/st/wrp_psp/31F0.c
@@ -1,7 +1,0 @@
-#include "common.h"
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/31F0", Update);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/31F0", UpdateStageEntities);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/31F0", func_psp_0923AD68);
