--- src/st/wrp_psp/3C70.c
@@ -1,3 +1,0 @@
-#include "common.h"
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/3C70", func_psp_0923B2F0);
