--- src/st/wrp_psp/5AB0.c
@@ -1,9 +1,0 @@
-#include "common.h"
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/5AB0", DestroyCurrentEntity);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/5AB0", EntityPrizeDrop);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/5AB0", EntityExplosion);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/5AB0", BlinkItem);
