--- src/st/wrp_psp/6BF0.c
@@ -1,5 +1,0 @@
-#include "common.h"
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/6BF0", EntityEquipItemDrop);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/6BF0", EntityRelicOrb);
