--- src/st/wrp_psp/80.c
@@ -1,5 +1,0 @@
-#include "../wrp/wrp.h"
-
-#include "../create_entity_from_layout.h"
-
-#include "../create_entity_in_range.h"
