--- src/st/wrp_psp/8580.c
@@ -1,15 +1,0 @@
-#include "common.h"
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/8580", EntityHeartDrop);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/8580", EntityMessageBox);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/8580", func_8018F510);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/8580", func_8018F838);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/8580", EntityUnkId15);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/8580", EntityIntenseExplosion);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/8580", func_801916C4);
