--- src/st/wrp_psp/95D0.c
@@ -1,7 +1,0 @@
-#include "../wrp/wrp.h"
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/95D0", TestCollisions);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/95D0", EntityNumericDamage);
-
-INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/95D0", BottomCornerText);
