+++ src/st/wrp_psp/collision.c
@@ -1,0 +1,5 @@
+#include "../wrp/wrp.h"
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/collision", TestCollisions);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/collision", EntityNumericDamage);
