+++ src/st/wrp_psp/e_misc.c
@@ -1,0 +1,17 @@
+#include "../wrp/wrp.h"
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_misc", EntityRelicOrb);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_misc", EntityHeartDrop);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_misc", EntityMessageBox);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_misc", func_8018F510);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_misc", func_8018F838);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_misc", EntityUnkId15);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_misc", EntityIntenseExplosion);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_misc", func_801916C4);
