+++ src/st/wrp_psp/e_particles.c
@@ -1,0 +1,9 @@
+#include "../wrp/wrp.h"
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_particles", func_psp_0923AD68);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_particles", func_psp_0923B2F0);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_particles", EntitySoulStealOrb);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_particles", EntityEnemyBlood);
