+++ src/st/wrp_psp/e_room_fg.c
@@ -22,13 +22,3 @@ void EntityRoomForeground(Entity* entity) {
     }
     AnimateEntity(objInit->unk10, entity);
 }
-
-#include "../../destroy_entity.h"
-
-void PreventEntityFromRespawning(Entity* entity) {
-    if (entity->entityRoomIndex) {
-        u16 index = entity->entityRoomIndex - 1 >> 5;
-        g_unkGraphicsStruct.D_80097428[index] |=
-            1 << ((entity->entityRoomIndex - 1) & 0x1F);
-    }
-}
