+++ src/st/wrp_psp/e_stage_name.c
@@ -1,0 +1,3 @@
+#include "../wrp/wrp.h"
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/e_stage_name", EntityStageNamePopup);
