+++ src/st/wrp_psp/popup.c
@@ -1,0 +1,3 @@
+#include "../wrp/wrp.h"
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/popup", BottomCornerText);
