+++ src/st/wrp_psp/st_update.c
@@ -1,0 +1,5 @@
+#include "../wrp/wrp.h"
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/st_update", Update);
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/st_update", UpdateStageEntities);
