+++ src/st/wrp_psp/stage_name_new.c
@@ -1,0 +1,3 @@
+#include "../wrp/wrp.h"
+
+INCLUDE_ASM("st/wrp_psp/psp/wrp_psp/stage_name_new", func_psp_09244020);
